+++ src/main/kotlin/org/soneira/savings/domain/vo/id/CategoryId.kt
@@ -1,5 +1,6 @@
 package org.soneira.savings.domain.vo.id
 
 import org.soneira.savings.domain.common.vo.BaseId
+import java.math.BigInteger
 
-class CategoryId(value: Int) : BaseId<Int>(value)
+class CategoryId(value: BigInteger) : BaseId<BigInteger>(value)
