+++ src/main/kotlin/org/soneira/savings/infrastructure/persistence/mongo/document/MovementDocument.kt
@@ -5,6 +5,7 @@ import org.springframework.data.mongodb.core.mapping.Document
 import org.springframework.data.mongodb.core.mapping.Field
 import org.springframework.data.mongodb.core.mapping.FieldType
 import java.math.BigDecimal
+import java.math.BigInteger
 import java.time.LocalDate
 
 @Document("movements")
@@ -13,7 +14,7 @@ data class MovementDocument(
     val description: String,
     @Field(targetType = FieldType.DECIMAL128) val amount: BigDecimal,
     val order: Int,
-    val subcategory: Int,
+    val subcategory: BigInteger,
     var comment: String,
     @Field(targetType = FieldType.DECIMAL128) val balance: BigDecimal
 ) {
@@ -23,7 +24,7 @@ data class MovementDocument(
 
     constructor(
         id: String, periodId: String, operationDate: LocalDate, description: String, amount: BigDecimal, order: Int,
-        subcategory: Int, comment: String, balance: BigDecimal
+        subcategory: BigInteger, comment: String, balance: BigDecimal
     ) :
             this(operationDate, description, amount, order, subcategory, comment, balance) {
         this.id = id
