+++ src/main/kotlin/org/soneira/savings/infrastructure/persistence/mongo/document/SettingsDocument.kt
@@ -1,12 +1,13 @@
 package org.soneira.savings.infrastructure.persistence.mongo.document
 
 import org.soneira.savings.domain.vo.PeriodStrategyType
+import java.math.BigInteger
 
 data class SettingsDocument(
     val periodStrategyType: PeriodStrategyType,
-    val periodDefiner: Int,
+    val periodDefiner: BigInteger,
     val monthStartBoundary: Int,
     val monthEndBoundary: Int,
-    val subcategoriesNotCountable: List<Int>,
+    val subcategoriesNotCountable: List<BigInteger>,
     val fileParserSettings: FileParserSettingsDocument
 )
